Rename PaddleOCR recognizer module
